services/geo_pipeline/requirements.txt
fastapi
uvicorn[standard]
redis
pydantic
Pillow # For image metadata extraction
async-timeout
python-multipart # For FastAPI File Uploads
httpx # For calling other services
